refactoring and load with scroll
